migration updated

+++ database/migrations/2024_06_09_215909_create_room_prices_table.php
@@ -4,20 +4,32 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
-    public function up(): void
+    public function up() : void
     {
         Schema::create('room_prices', function (Blueprint $table) {
             $table->id();
-            $table->double('price')->default(0);
+            $table->integer('price')->default(0);
             $table->string('note')->nullable();
-            $table->boolean('is_activated')->default(0)->nullable('0 means not activate');//Only 1 active price at time for rooms
-            $table->foreignId('price_type_id')->nullable()->constrained()->onDelete('Set Null')->onUpdate('No Action');
-            $table->foreignId('room_id')->constrained()->onDelete('cascade')->onUpdate('No Action');
+
+            $table->boolean('is_activated')
+                ->default(0)
+                ->nullable('0 means not activate');//Only 1 active price at time for rooms
+
+            $table->foreignId('price_type_id')
+                ->nullable()
+                ->constrained()
+                ->onDelete('Set Null')
+                ->onUpdate('No Action');
+
+            $table->foreignId('room_id')
+                ->constrained()
+                ->onDelete('cascade')
+                ->onUpdate('No Action');
+
             $table->softDeletes();
             $table->timestamps();
         });
@@ -26,7 +38,7 @@ return new class extends Migration
     /**
      * Reverse the migrations.
      */
-    public function down(): void
+    public function down() : void
     {
         Schema::dropIfExists('room_prices');
     }
